Moved train_heart_predictor script to the resources folder in training-api. Also adjusted train script so it works with app.py to handle the necessary HTTP requests.
