--- train_heart_predictor.py
@@ -1,118 +1,0 @@
-import os
-os.environ['TF_ENABLE_ONEDNN_OPTS'] = '0'
-
-import pandas as pd
-import numpy as np
-from sklearn.model_selection import train_test_split
-from sklearn.preprocessing import StandardScaler
-from sklearn.metrics import confusion_matrix, classification_report
-import matplotlib.pyplot as plt
-import seaborn as sns
-import tensorflow as tf
-from tensorflow.keras.models import Sequential
-from tensorflow.keras.layers import Dense, Dropout
-from tensorflow.keras.callbacks import EarlyStopping, LearningRateScheduler
-from tensorflow.keras.regularizers import l2
-
-# Load dataset
-dataset_path = 'heart.csv'
-if not os.path.exists(dataset_path):
-    raise FileNotFoundError(f"Dataset not found at {dataset_path}")
-
-data = pd.read_csv(dataset_path)
-
-# Features and target
-X = data.iloc[:, :-1]
-y = data.iloc[:, -1]
-
-# Split the data into train (60%), validation (20%), and test (20%) sets
-X_train, X_temp, y_train, y_temp = train_test_split(X, y, test_size=0.4, random_state=42)
-X_val, X_test, y_val, y_test = train_test_split(X_temp, y_temp, test_size=0.5, random_state=42)
-
-# Standardize the data
-scaler = StandardScaler()
-X_train_scaled = scaler.fit_transform(X_train)
-X_val_scaled = scaler.transform(X_val)
-X_test_scaled = scaler.transform(X_test)
-
-# Learning rate scheduler function
-def lr_scheduler(epoch, lr):
-    if epoch > 10:
-        lr = lr * 0.9  # Reduce learning rate by 10% after 10 epochs
-    return lr
-
-# Build a neural network model
-model = Sequential()
-model.add(Dense(64, input_dim=X_train_scaled.shape[1], activation='relu', kernel_regularizer=l2(0.001)))  # L2 regularization
-model.add(Dropout(0.3))  # Dropout layer to prevent overfitting
-model.add(Dense(32, activation='relu', kernel_regularizer=l2(0.001)))
-model.add(Dropout(0.3))
-model.add(Dense(1, activation='sigmoid'))  # Output layer for binary classification
-
-# Compile the model
-model.compile(loss='binary_crossentropy',
-              optimizer='adam',
-              metrics=['accuracy'])
-
-# Early stopping and learning rate scheduler callbacks
-early_stopping = EarlyStopping(monitor='val_loss', patience=5, restore_best_weights=True)
-lr_scheduler_cb = LearningRateScheduler(lr_scheduler)
-
-# Train the model using validation data
-history = model.fit(X_train_scaled, y_train,
-                    epochs=100,
-                    batch_size=32,
-                    validation_data=(X_val_scaled, y_val),
-                    callbacks=[early_stopping, lr_scheduler_cb],
-                    verbose=1)
-
-# Save the model
-model.save('heart_disease_model.h5')
-
-# Evaluate the model on the test set (final evaluation)
-test_loss, test_acc = model.evaluate(X_test_scaled, y_test)
-print(f"Test accuracy: {test_acc:.4f}")
-
-# Predictions
-y_pred = (model.predict(X_test_scaled) > 0.5).astype(int)
-
-# Confusion Matrix
-conf_matrix = confusion_matrix(y_test, y_pred)
-plt.figure(figsize=(8, 6))
-sns.heatmap(conf_matrix, annot=True, fmt="d", cmap="Blues")
-plt.title("Confusion Matrix")
-plt.ylabel("Actual Label")
-plt.xlabel("Predicted Label")
-plt.show()
-
-# Classification Report
-print(classification_report(y_test, y_pred))
-
-# Plot training & validation accuracy and loss to detect overfitting/underfitting
-plt.figure(figsize=(14, 6))
-plt.subplot(1, 2, 1)
-plt.plot(history.history['accuracy'], label='Train Accuracy')
-plt.plot(history.history['val_accuracy'], label='Validation Accuracy')
-plt.title('Accuracy Over Epochs')
-plt.xlabel('Epochs')
-plt.ylabel('Accuracy')
-plt.legend()
-
-plt.subplot(1, 2, 2)
-plt.plot(history.history['loss'], label='Train Loss')
-plt.plot(history.history['val_loss'], label='Validation Loss')
-plt.title('Loss Over Epochs')
-plt.xlabel('Epochs')
-plt.ylabel('Loss')
-plt.legend()
-
-plt.show()
-
-# Plot the learning rate over epochs to verify the effect of lr_scheduler
-lrs = [lr_scheduler_cb.lr_schedule(epoch, 0.001) for epoch in range(len(history.history['loss']))]
-plt.figure(figsize=(6, 4))
-plt.plot(lrs)
-plt.title('Learning Rate Over Epochs')
-plt.xlabel('Epochs')
-plt.ylabel('Learning Rate')
-plt.show()
