+++ training-api/.dockerignore
@@ -1,0 +1,3 @@
+.git
+.gitignore
+Dockerfile*
