+++ training-api/requirements.txt
@@ -1,0 +1,8 @@
+setuptools
+flask
+pandas
+keras
+tensorflow
+numpy==1.26.4
+h5py
+six
